+++ Examples/IcmpFileTransfer/IcmpFileTransfer-pitcher.cpp
@@ -112,7 +112,7 @@ static void waitForFileTransferStart(RawPacket* rawPacket, PcapLiveDevice* dev, 
 
 	// verify the source IP is the catcher's IP and the dest IP is the pitcher's IP
 	IPv4Layer* ip4Layer = parsedPacket.getLayerOfType<IPv4Layer>();
-	if (ip4Layer->getSrcIpAddress().toInt() != icmpFTStart->catcherIPAddr.toUInt() || ip4Layer->getDstIpAddress().toInt() != icmpFTStart->pitcherIPAddr.toUInt()) // // TODO: remove toUInt()/toInt() when migration has done
+	if (ip4Layer->getSrcIpAddress() != icmpFTStart->catcherIPAddr || ip4Layer->getDstIpAddress() != icmpFTStart->pitcherIPAddr)
 		return;
 
 	// extract the message type in the ICMP reply timestamp field and check if it's  ICMP_FT_START
@@ -157,7 +157,7 @@ static void getFileContent(RawPacket* rawPacket, PcapLiveDevice* dev, void* icmp
 
 	// verify the source IP is the catcher's IP and the dest IP is the pitcher's IP
 	IPv4Layer* ip4Layer = parsedPacket.getLayerOfType<IPv4Layer>();
-	if (ip4Layer->getSrcIpAddress().toInt() != icmpFileContentData->catcherIPAddr.toUInt() || ip4Layer->getDstIpAddress().toInt() != icmpFileContentData->pitcherIPAddr.toUInt()) // // TODO: remove toUInt()/toInt() when migration has done
+	if (ip4Layer->getSrcIpAddress() != icmpFileContentData->catcherIPAddr || ip4Layer->getDstIpAddress() != icmpFileContentData->pitcherIPAddr)
 		return;
 
 	// extract the message type from the ICMP reply timestamp field
@@ -385,7 +385,7 @@ static bool waitForFileTransferStartAck(RawPacket* rawPacket, PcapLiveDevice* de
 
 	// verify the source IP is the catcher's IP and the dest IP is the pitcher's IP
 	IPv4Layer* ip4Layer = parsedPacket.getLayerOfType<IPv4Layer>();
-	if (ip4Layer->getSrcIpAddress().toInt() != icmpData->catcherIPAddr.toUInt() || ip4Layer->getDstIpAddress().toInt() != icmpData->pitcherIPAddr.toUInt()) // TODO: remove toUInt()/toInt() when migration has done
+	if (ip4Layer->getSrcIpAddress() != icmpData->catcherIPAddr || ip4Layer->getDstIpAddress() != icmpData->pitcherIPAddr)
 		return false;
 
 	// verify the message type is ICMP_FT_ACK
@@ -404,7 +404,7 @@ static bool waitForFileTransferStartAck(RawPacket* rawPacket, PcapLiveDevice* de
 void sendFile(std::string filePath, pcpp::experimental::IPv4Address pitcherIP, pcpp::experimental::IPv4Address catcherIP, size_t blockSize, int packetPerSec)
 {
 	// identify the interface to listen and send packets to
-	PcapLiveDevice* dev = PcapLiveDeviceList::getInstance().getPcapLiveDeviceByIp(pitcherIP.toUInt()); // // TODO: remove toUInt() when migration has done
+	PcapLiveDevice* dev = PcapLiveDeviceList::getInstance().getPcapLiveDeviceByIp(pitcherIP.toUInt()); // TODO: remove toUInt() when migration has done
 	if (dev == NULL)
 		EXIT_WITH_ERROR("Cannot find network interface with IP '%s'", pitcherIP.toString().c_str());
